selection/getRangeAt.html

<!DOCTYPE html>
<title>The getRangeAt method</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<link rel="help" href="https://w3c.github.io/selection-api/#dom-selection-getrangeat">
<script>
const sel = getSelection();

test(function() {
  var range = document.createRange();
  sel.addRange(range);
  assert_throws_dom("INDEX_SIZE_ERR", function() { sel.getRangeAt(-1); })
  assert_throws_dom("INDEX_SIZE_ERR", function() { sel.getRangeAt(1); })
}, "Only supports index 0");

test(() => {
  assert_equals(sel.getRangeAt(0), sel.getRangeAt(0));
}, "Returns the same range object");
</script>
